% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiAddRowsToTable.R
\name{pbiAddRowsToTable}
\alias{pbiAddRowsToTable}
\title{Add rows to a table.}
\usage{
pbiAddRowsToTable(df, guid, tableName, truncate = FALSE)
}
\arguments{
\item{df}{The data frame you want to insert.}

\item{guid}{Dataset Id.}

\item{tableName}{The friendly name of the table. The URL will be encoded.}

\item{truncate}{When TRUE, calls pbiTruncateDataset to clean the table before loading.}
}
\description{
Assumes that the schema of the data frame matches the table, it will try and insert the rows.
}
\examples{
\dontrun{pbiAddRowsToTable(iris, guid = "abc", tableName = "Iris")}
}
