% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiAuthenticate.R
\name{pbiAuthenticate}
\alias{pbiAuthenticate}
\title{Authenticate your Power BI API session.}
\usage{
pbiAuthenticate(appName, clientId, clientSecret)
}
\arguments{
\item{appName}{The name of the application.}

\item{clientId}{Application ID.}

\item{clientSecret}{Client secret.}

\item{cache}{When TRUE, a .httr-oauth file will be created in the working directory which be reused for further authentications. This will mean you only have to enter your password on the Windows login once - an option for scripting processes. When FALSE the user will have to authenticate with a password each time.}
}
\description{
Supply your application name, client ID and client secret. You will need to register an application here: https://powerbi.microsoft.com/en-us/documentation/powerbi-developer-register-a-web-app/
}
\examples{
pbiAuthenticate(appName = "abc", clientId = "def", clientSecret = "ghi")
}
