% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiCreateDataset.R
\name{pbiCreateDataset}
\alias{pbiCreateDataset}
\title{Create a new dataset.}
\usage{
pbiCreateDataset(name = NULL, duplicate = FALSE, groupId = NULL,
  tableSchema = list(), measureSchema = list(),
  relationshipSchema = list(), FIFO = "none", verbose = FALSE)
}
\arguments{
\item{name}{The name of the new dataset.}

\item{duplicate}{When TRUE, it will create a separate dataset with the same name as another dataset.}

\item{groupId}{Specify which group to create it to. Use pbiListAllUserGroups to find the ids.}

\item{tableSchema}{A list object containing relevant column metadata. You may use pbiGenerateTableSchema to help you.}

\item{measureSchema}{A list object containing relevant information about measures.}

\item{relationshipSchema}{A list object containing relevant relationship metadata.}

\item{FIFO}{Default retention policy, or basic "first-in-first-out". Dataset will store between 200-210k rows and remove the oldest rows when the storage limit has been reached. Consider this option for streaming datasets. Options are 'none' or 'basicFIFO'.}
}
\value{
HTTP response object.
}
\description{
Requires the correct metadata information to be passed to the function in the form of lists and named variables. Relationships, measures and columns can be handled.
}
