% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiCreateDatasetFromDataFrame.R
\name{pbiCreateDatasetFromDataFrame}
\alias{pbiCreateDatasetFromDataFrame}
\title{Create a new dataset from a data frame.}
\usage{
pbiCreateDatasetFromDataFrame(name, df, tableName, duplicate = FALSE,
  addRows = TRUE, datasetType = "Push", FIFO = "none", verbose = FALSE)
}
\arguments{
\item{name}{The name of the new dataset.}

\item{df}{A data frame to determine the table schema. Uses pbiGenerateTableSchema to infer the data types and creates a JSON object.}

\item{tableName}{An character vector for a table name (max 1 table created per call)}

\item{duplicate}{When TRUE, it will create a separate dataset with the same name as another dataset.}

\item{addRows}{When TRUE, it will create the schema and data. When FALSE, just the schema will be created.}

\item{datasetType}{What kind of dataset do you want to use? Use pbiDatasetTypes to see the available options.}

\item{FIFO}{Default retention policy, or basic "first-in-first-out". Dataset will store between 200-210k rows and remove the oldest rows when the storage limit has been reached. Consider this option for streaming datasets. Options are 'none' or 'basicFIFO'.}

\item{verbose}{If TRUE, return extra information about the endpoint.}
}
\value{
HTTP response object.
}
\description{
Requires the correct metadata information to be passed to the function in the form of lists and named variables. Relationships, measures and columns can be handled.
}
\examples{
\dontrun{pbiCreateDatasetFromDataFrame("New Dataset", df = cars, tableName = "Cars")}
}
