% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiDatasetTypes.R
\name{pbiDatasetTypes}
\alias{pbiDatasetTypes}
\title{Supported dataset types.}
\usage{
pbiDatasetTypes(simple = TRUE, verbose = FALSE)
}
\value{
List
}
\description{
Dataset types.
}
\examples{
pbiDatasetTypes()
}
