% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiDataTypes.R
\name{pbiDataTypes}
\alias{pbiDataTypes}
\title{Supported data types.}
\usage{
pbiDataTypes()
}
\arguments{
\item{simple}{If TRUE, simply return a character vector of the names. Else, return a detailed data frame with restrictions.}

\item{verbose}{If TRUE, it will report information about Power BI data types to the user.}
}
\value{
Data frame.
}
\description{
A data frame contain supported data types and restrictions, sourced from the REST API reference.
}
\examples{
pbiDataTypes()
}
