% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiDeleteDataset.R
\name{pbiDeleteDataset}
\alias{pbiDeleteDataset}
\title{Delete a dataset.}
\usage{
pbiDeleteDataset(guid)
}
\arguments{
\item{guid}{A character string of a guid for a single dataset.}
}
\value{
cat
}
\description{
Remove the dataset entirely, not just truncate the rows of a table.
}
\examples{
\dontrun{pbiDeleteDataset(id = "as8sdfasdfjkha")}
}
