% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiGenerateTableSchema.R
\name{pbiGenerateTableSchema}
\alias{pbiGenerateTableSchema}
\title{Generate a table schema.}
\usage{
pbiGenerateTableSchema(df, tableName = NULL, debug = FALSE)
}
\arguments{
\item{df}{Data frame - the data used to infer the table schema.}

\item{tableName}{A character string. When is not NULL, it will override the default of using the variable name as the table name.}

\item{debug}{When TRUE, it will return a list object converted from the JSON so that you can see what the R object should look like.}
}
\value{
List object or JSON to be passed into the body of a request.
}
\description{
Use a data frame to infer data types and produce a list object that can be converted to compatible JSON. See here: https://msdn.microsoft.com/en-us/library/mt203560.aspx
}
\examples{
pbiGenerateTableSchema(cats, debug = TRUE)
pbiDataTypes() # Use to reference accepted data types.
}
