% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiListAllDashboards.R
\name{pbiListAllDashboards}
\alias{pbiListAllDashboards}
\title{List all dashboards.}
\usage{
pbiListAllDashboards(toDf = TRUE)
}
\arguments{
\item{toDf}{When TRUE, a tabular format is returned. When FALSE, more detailed content is returned as JSON.}
}
\value{
Data frame or JSON.
}
\description{
Return a data frame containing basic information about all of the dashboards on your Power BI web service. Requires authentication before it can be executed.
}
\examples{
\dontrun{pbiListAllDashboards(toDf = FALSE)}
}
