% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiListAllReports.R
\name{pbiListAllReports}
\alias{pbiListAllReports}
\title{List all reports.}
\usage{
pbiListAllReports(toDf = TRUE)
}
\arguments{
\item{toDf}{When TRUE, a tabular format is returned. When FALSE, more detailed content is returned as JSON.}
}
\value{
Data frame.
}
\description{
Return a data frame containing basic information about all of the reports on your Power BI web service. Requires authentication before it can be executed.
}
\examples{
pbiListAllReports()
}
