% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiListAllTables.R
\name{pbiListAllTables}
\alias{pbiListAllTables}
\title{List all tables (from all datasets)}
\usage{
pbiListAllTables(id = NULL, allDatasets = FALSE)
}
\arguments{
\item{id}{A character string or vector containing guids. Hint: You can use pbiListAllDatasets to help you.}

\item{all}{When TRUE, it will override the value of id and proceed to find all datasets.}
}
\value{
Data frame.
}
\description{
Return all tables from a single, multiple or from all datasets.
}
\examples{
pbiListAllTables(id = "as876fdsjds")
}
