% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiListAllTiles.R
\name{pbiListAllTiles}
\alias{pbiListAllTiles}
\title{List all tiles (from all dashboards)}
\usage{
pbiListAllTiles(id = NULL, allDashboards = FALSE)
}
\arguments{
\item{id}{A character string or vector containing guids. Hint: You can use pbiListAllDashboards to help you.}

\item{all}{When TRUE, it will override the value of id and proceed to find all dashboards.}
}
\value{
Data frame.
}
\description{
Return all tiles from a single, multiple or from all dashboards.
}
\examples{
pbiListAllTiles(all = TRUE)
}
