% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiListAllUserGroups.R
\name{pbiListAllUserGroups}
\alias{pbiListAllUserGroups}
\title{List all imported files.}
\usage{
pbiListAllUserGroups(toDf = TRUE)
}
\arguments{
\item{toDf}{When TRUE, a tabular format is returned. When FALSE, more detailed content is returned as JSON.}
}
\value{
Data frame.
}
\description{
Return a data frame containing information about the groups associated with the user.
}
\examples{
\dontrun{pbiListAllUserGroups()}
}
