% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiQueryBuilder.R
\name{pbiQueryBuilder}
\alias{pbiQueryBuilder}
\title{Build simple REST API queries.}
\usage{
pbiQueryBuilder(endpoint, guid = NULL, object = NULL, suffix = NULL,
  method = "GET", version = "v1.0", printCall = FALSE,
  getContent = TRUE)
}
\arguments{
\item{endpoint}{Which endpoint of the API to query.}

\item{method}{Default is GET.}

\item{version}{The API version.}

\item{printCall}{Print the URL of the API call used for debugging.}

\item{getContent}{If TRUE, extract JSON from the response object.}
}
\value{
HTTP response object.
}
\description{
An internal function that acts as a skeleton for building API queries for other functions. This will only work when you have authenticated first.
}
\examples{
pbiQueryBuilder(endpoint = "datasets")
}
