% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiTruncateDataset.R
\name{pbiTruncateDataset}
\alias{pbiTruncateDataset}
\title{Truncate a table.}
\usage{
pbiTruncateDataset(guid, tableName)
}
\arguments{
\item{guid}{A single character string containing a guid.}

\item{tableName}{A single character string containing an pre-encoded friendly table name.}
}
\value{
HTTP response.
}
\description{
Remove all rows from a table within a dataset. Hint: You can use pbiListAllTables to do this across multiple tables.
}
\examples{
pbiTruncateDataset(guid = "aasdafs876fdsjds", tableName = "June Sales")
}
