% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiUpdateTableSchema.R
\name{pbiUpdateTableSchema}
\alias{pbiUpdateTableSchema}
\title{Update an existing table schema.}
\usage{
pbiUpdateTableSchema(df, guid, tableName)
}
\arguments{
\item{df}{A data frame off of which to build the table schema. It can be empty or contain rows, so long as the types are correct.}

\item{guid}{Dataset id. Use pbiListAllDatasets.}

\item{tableName}{Friendly table name - see pbiListAllTables for help.}
}
\value{
HTTP response object.
}
\description{
Requires dataset id and table name and a data.frame. Passes a table schema to the body of the request - uses pbiGenerateTableSchema to dynamically create a table schema off of a data frame, which you can use to give you an idea about what the R object should look like if you want to create a data frame yourself.
}
\examples{
pbiUpdateTableSchema(iris, guid = "sdjfalsdkfj2", tableName = "Friendly Name")
pbiGenerateTableSchema(iris)
}
