% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiAuthenticate.R
\name{pbiAuthenticate}
\alias{pbiAuthenticate}
\title{Authenticate your Power BI API session.}
\usage{
pbiAuthenticate(appName, clientId, clientSecret)
}
\arguments{
\item{appName}{The name of the application.}

\item{clientId}{Application ID.}

\item{clientSecret}{Client secret.}
}
\description{
Supply your application name, client ID and client secret. You will need to register an application here: https://powerbi.microsoft.com/en-us/documentation/powerbi-developer-register-a-web-app/
}
\examples{
pbiAuthenticate(appName = "abc", clientId = "def", clientSecret = "ghi")
}
